+++ app/_services/tokenApproval.ts
@@ -10,7 +10,7 @@ export const hasEnoughTokenApproval = async (
   requesterCustodyAddress: `0x${string}`
 ) => {
   const deployment =
-    yamlData.deployments[currentState.deploymentOption.deployment];
+    yamlData.deployments[currentState.deploymentOption?.deployment || ""];
   const order = yamlData.orders[deployment.order];
 
   const orderBook = yamlData.orderbooks[order.orderbook];
