+++ app/my-strategies/page.tsx
@@ -3,35 +3,10 @@
 import { useQuery } from "@tanstack/react-query";
 import { getOrders } from "../_queries/getOrders";
 import { useAccount } from "wagmi";
-import { useEffect } from "react";
 import { Spinner, Table } from "flowbite-react";
 import { formatTimestampSecondsAsLocal } from "../_services/dates";
-import { formatUnits } from "viem";
 import { useRouter } from "next/navigation";
-import { WithdrawalModal } from "../_components/WithdrawalModal";
-
-export function TokenAndBalance({
-  input,
-  withdraw,
-}: {
-  input: any;
-  withdraw?: boolean;
-}) {
-  return (
-    <div className="flex border rounded-xl p-2 gap-x-3 items-center justify-between">
-      <div className="flex flex-col gap-y-1">
-        <div>{withdraw ? input.token.name : input.token.symbol}</div>
-        <div className="text-gray-500">
-          Balance:{" "}
-          {Number(
-            Number(formatUnits(input.balance, input.token.decimals)).toFixed(8)
-          )}
-        </div>
-      </div>
-      {withdraw && <WithdrawalModal vault={input} />}
-    </div>
-  );
-}
+import { TokenAndBalance } from "../_components/TokenAndBalance";
 
 export default function MyStrategies() {
   const router = useRouter();
